Fix return value
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/branches/libsidplayfp-experimental@2747 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/builders/residfp-builder/residfp/SID.h
@@ -118,7 +118,7 @@ private:
      *
      * @return the output sample
      */
-    int output() const;
+    float output() const;
 
     /**
      * Calculate the numebr of cycles according to current parameters
@@ -298,7 +298,7 @@ void SID::ageBusValue(unsigned int n)
 }
 
 RESID_INLINE
-int SID::output() const
+float SID::output() const
 {
     const float v1 = voice[0]->output(voice[2]->wave()) / (65536.f * 16.f);
     const float v2 = voice[1]->output(voice[0]->wave()) / (65536.f * 16.f);
